Fixup - remove JsonIgnore from Metadata.Table.Model property
The `[JsonIgnore]` attribute was removed from the `Model` property in the `Dax.Metadata` namespace. Comments were added to explain the history of this change, noting that the attribute was initially introduced in VAPX 1.11-preview2 but was commented out to avoid breaking changes during deserialization with earlier library versions. The comments also reference a related pull request and suggest the possibility of uncommenting the attribute in future versions.